Cleanup of spatial module and added ops.
