+++ spatial/shared/src/main/scala/io/youi/spatial/ops/Matrix3Modifier.scala
@@ -1,6 +1,6 @@
-package io.youi.spatial.prototype.ops
+package io.youi.spatial.ops
 
-import io.youi.spatial.prototype.Matrix3
+import io.youi.spatial.Matrix3
 
 trait Matrix3Modifier {
   type Mod = (Double, Double, Double, Double, Double, Double, Double, Double, Double) => Matrix3
